Fixed problem with BINARY_BUNDLE_NAME that was a result of recent -version changes. For a J2ME_CLASSLIB=personal build, it was producing the following:
phoneME_Advanced-phoneme_advanced_mr2_bXX-linux_x86_generic-bin-rev1472

I changed it to:

Personal_Profile-phoneme_advanced_mr2_bXX-linux_x86_generic-bin-rev1472

The format is basically:

<topmost_profile_name>-<release>_<build>-<platform>-bin-<svn-revision>

The <release> is what you use when you file bugs. <build> is also used for filing bugs, but at the moment we haven't decided how/if to properly convey this in builds off the svn trunk. This will probably all change in the near future as we work through versioning details, but for now I wanted to just fix the missing profile information.

Also, I fixed a long standing problem with the svn revision number suffix on BINARY_BUNDLE_DIRNAME. The revision number was in there twice. I was sure I fixed this once, but it looks like I only made this fix for the source bundle dirname.

Reviwed by Dean