Fixed CR 6520226 - Attempt to create DatagraSocket using non-local IPv6 address causes BindException
This CR is similar to JavaSE bug 6206527. Linux should allow the application to use a scope_id of zero (broken in kernel 2.6 because it does not ignore the scope_id in the bind() call. It presumably ignored prior to 2.6) which would force the kernel to try an figure out which interface to use (in most cases it would be possible to determine it).

Since it does not do this, we need to use a different method for determining the scope_id when binding to a local address. In this case, we will search /proc/net/if_inet6 instead of /proc/net/ipv6_route.

Testing:
1. Ran CDC and FP TCK java_net tests
2. Unit tests with the bug report