Prototype changes for StartMIDletMonitor common for MVM and SVM:
- StartMIDletEntry is decoupled from StartMIDletMonitor
- MVMStartMIDletEntry extends StartMIDletEntry with Isolates specific
- Proper factories for StartMIDletEntry creation are designed and
  parameterize StartMIDletMonitor
- Both AmsUtil versions (SVM/MVM) use StartMIDletMonitor

The unsolved problem is that StartMIDletMonitor is not aware of 1st started 
MIDlet in VM (AMS in MVM, or user MIDlet in SVM on VM cycling). That makes 
the monitor useless for SVM, since StartMIDletMonitor can't detect repeated 
execution of the 1st started MIDlet in VM. It is possible to solve the problem 
moving StartMIDletMonitor initialization from AmsUtil to system scope, and 
obligatory registration within it of any MIDlet being started. It will 
introduce more and more changes, that's dangerous on the code freeze threshold.