6631002: doing two parallel builds fails if done in same build directory but with different subdirs
I moved the temp files to the generated directory. I also no longer delete .DefaultLocaleList.java in the clean rule since it would be rare for it to be left around, and if it is it will be hidden in the generated directory.

I also did one other minor change. There was a $(error ...) string with a single quote in it. This messes up the syntax highlighter in xemacs. It thinks everything after it until the next single quote is part of a quoted string, so I removed the single quote. 

Reviewed by Mark
Did parellel test builds.