6819090: 3 CDC TCK tests fail when using the split verifier
See 4817320, 6293528, and other related bugs for the instr_03801m1 failure. The backwards branch check has been removed from the split veriier.

For the tableswitch and lookswitch failure, the problem is that the test case has non-zero values in the pad bytes, which is not allowed. The split verifier now detects and rejects these classes.

Reviewed by Dean and Bill.